+++ src/packages.lisp
@@ -4,7 +4,8 @@
   (:shadow #:bit #:and)
   (:export ;; Public API
            #:execute #:6502-step #:asm #:disasm #:disasm-to-str #:disasm-to-list
-           #:get-byte #:get-word #:get-range #:*cpu* #:cpu #:nmi #:reset
+           #:current-instruction #:get-byte #:get-word #:get-range #:*cpu* #:cpu
+           #:nmi #:reset
            ;; CPU struct
            #:make-cpu #:cpu-ar #:cpu-xr #:cpu-yr #:cpu-sr #:cpu-sp #:cpu-pc #:cpu-cc #:u8
            ;; Addr modes
@@ -17,6 +18,7 @@
   (:documentation "Homepage: <a href=\"http://github.com/redline6561/cl-6502\">Github</a>")
   (:use :cl)
   (:import-from :6502 #:execute #:6502-step #:asm #:disasm #:disasm-to-str #:disasm-to-list
-                      #:get-byte #:get-word #:get-range #:*cpu* #:cpu #:nmi #:reset)
+                #:current-instruction #:get-byte #:get-word #:get-range #:*cpu* #:cpu #:nmi #:reset)
   (:export #:execute #:6502-step #:asm #:disasm #:disasm-to-str #:disasm-to-list
-           #:get-byte #:get-word #:get-range #:*cpu* #:cpu #:nmi #:reset))
+           #:current-instruction #:get-byte #:get-word #:get-range #:*cpu* #:cpu
+           #:nmi #:reset))